added guard rails
